chore: Update PR template
The "Description" header is removed so that GitHub places the
commit message in the proper location. Otherwise, it has to be cut
and pasted below the header manually.

The checklist is simplified - the first point was misleading, since
changing the chart README.md *would* need a version bump. Also,
a typo is fixed in the second point.